// SDKs/onsched.com/setup/v1/typescript/src/sdk/models/shared/addressinputmodel.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class AddressInputModel extends SpeakeasyBase {
  @Metadata({ data: "json, name=addressLine1" })
  addressLine1?: string;

  @Metadata({ data: "json, name=addressLine2" })
  addressLine2?: string;

  @Metadata({ data: "json, name=city" })
  city?: string;

  @Metadata({ data: "json, name=country" })
  country?: string;

  @Metadata({ data: "json, name=postalCode" })
  postalCode?: string;

  @Metadata({ data: "json, name=state" })
  state?: string;
}